Change qHash() to work with size_t instead of uint
This is required, so that QHash and QSet can hold more
than 2^32 items on 64 bit platforms.

The actual hashing functions for strings are still 32bit, this will
be changed in a follow-up commit.

+++ examples/embedded/lightmaps/slippymap.cpp
@@ -53,7 +53,7 @@
 #include "slippymap.h"
 #include "qmath.h"
 
-uint qHash(const QPoint& p)
+size_t qHash(const QPoint& p)
 {
     return p.x() * 17 ^ p.y();
 }
