+++ src/corelib/serialization/qcborarray.h
@@ -294,7 +294,7 @@ inline QCborArray QCborValueRef::toArray(const QCborArray &a) const
     return concrete().toArray(a);
 }
 
-Q_CORE_EXPORT uint qHash(const QCborArray &array, uint seed = 0);
+Q_CORE_EXPORT size_t qHash(const QCborArray &array, size_t seed = 0);
 
 #if !defined(QT_NO_DEBUG_STREAM)
 Q_CORE_EXPORT QDebug operator<<(QDebug, const QCborArray &a);
