+++ src/corelib/serialization/qcborcommon.h
@@ -138,12 +138,12 @@ QDataStream &operator<<(QDataStream &ds, QCborSimpleType st);
 QDataStream &operator>>(QDataStream &ds, QCborSimpleType &st);
 #endif
 
-inline uint qHash(QCborSimpleType tag, uint seed = 0)
+inline size_t qHash(QCborSimpleType tag, size_t seed = 0)
 {
     return qHash(quint8(tag), seed);
 }
 
-inline uint qHash(QCborTag tag, uint seed = 0)
+inline size_t qHash(QCborTag tag, size_t seed = 0)
 {
     return qHash(quint64(tag), seed);
 }
