+++ src/corelib/serialization/qjsonarray.h
@@ -252,7 +252,7 @@ private:
 
 Q_DECLARE_SHARED_NOT_MOVABLE_UNTIL_QT6(QJsonArray)
 
-Q_CORE_EXPORT uint qHash(const QJsonArray &array, uint seed = 0);
+Q_CORE_EXPORT size_t qHash(const QJsonArray &array, size_t seed = 0);
 
 #if !defined(QT_NO_DEBUG_STREAM) && !defined(QT_JSON_READONLY)
 Q_CORE_EXPORT QDebug operator<<(QDebug, const QJsonArray &);
