+++ src/gui/painting/qicc.cpp
@@ -136,7 +136,7 @@ enum class Tag : quint32 {
     aabg = IccTag('a', 'a', 'b', 'g'),
 };
 
-inline uint qHash(const Tag &key, uint seed = 0)
+inline size_t qHash(const Tag &key, size_t seed = 0)
 {
     return qHash(quint32(key), seed);
 }
